Colorsys
I tried changing my code to be able to use HSV values instead of RGB ones. I thought I got it to work, but there is some sort of error makes the color stay a similar hue. I've tried changing numbers to different values, but it all results in a reddish color in the image.